Revert "Change RawPeakElement's field type from double to float."
Revert below commit.
Commit 831bd767939daf375d365ec36ba44e58322c7ea4
Tag: BreakingChanges-SpectrumType

+++ CommonStandard/DataObj/RawMeasurement.cs
@@ -281,8 +281,8 @@ namespace CompMs.Common.DataObj {
 
     public struct RawPeakElement
     {
-        public float Mz;
-        public float Intensity;
+        public double Mz;
+        public double Intensity;
     }
 
     public struct Raw2DElement {
